// supabase/functions/interactions/types.d.ts
export type Body = {
  app_permissions: string;
  application_id: string;
  attachment_size_limit: number;
  authorizing_integration_owners: {
    "0": string;
  };
  channel: {
    flags: number;
    guild_id: string;
    icon_emoji: {
      id: any;
      name: string;
    };
    id: string;
    last_message_id: string;
    name: string;
    nsfw: boolean;
    parent_id: string;
    permissions: string;
    position: number;
    rate_limit_per_user: number;
    theme_color: any;
    topic: any;
    type: number;
  };
  channel_id: string;
  context: number;
  data: {
    component_type?: number;
    custom_id?: string;
    id: any;
    name?: string;
    type?: number;
    options?: Array<{
      name: string;
      type: number;
      value: string;
    }>;
  };
  entitlement_sku_ids: Array<any>;
  entitlements: Array<any>;
  guild: {
    features: Array<any>;
    id: string;
    locale: string;
  };
  guild_id: string;
  guild_locale: string;
  id: string;
  locale: string;
  member: {
    avatar: any;
    banner: any;
    collectibles: any;
    communication_disabled_until: any;
    deaf: boolean;
    display_name_styles: any;
    flags: number;
    joined_at: string;
    mute: boolean;
    nick: any;
    pending: boolean;
    permissions: string;
    premium_since: any;
    roles: Array<any>;
    unusual_dm_activity_until: any;
    user: {
      avatar: string;
      avatar_decoration_data: any;
      clan: {
        badge: string;
        identity_enabled: boolean;
        identity_guild_id: string;
        tag: string;
      };
      collectibles: any;
      discriminator: string;
      display_name_styles: any;
      global_name: any;
      id: string;
      primary_guild: {
        badge: string;
        identity_enabled: boolean;
        identity_guild_id: string;
        tag: string;
      };
      public_flags: number;
      username: string;
    };
  };
  message?: {
    application_id: string;
    attachments: Array<any>;
    author: {
      avatar: any;
      avatar_decoration_data: any;
      bot: boolean;
      clan: any;
      collectibles: any;
      discriminator: string;
      display_name_styles: any;
      global_name: any;
      id: string;
      primary_guild: any;
      public_flags: number;
      username: string;
    };
    channel_id: string;
    components: Array<{
      components: Array<{
        custom_id: string;
        disabled?: boolean;
        emoji: {
          name: string;
        };
        id: number;
        label: string;
        style: number;
        type: number;
      }>;
      id: number;
      type: number;
    }>;
    content: string;
    edited_timestamp: string;
    embeds: Array<{
      color: number;
      fields: Array<{
        inline: boolean;
        name: string;
        value: string;
      }>;
      id: string;
      title: string;
      type: string;
      url: string;
    }>;
    flags: number;
    id: string;
    interaction: {
      id: string;
      name: string;
      type: number;
      user: {
        avatar: string;
        avatar_decoration_data: any;
        clan: {
          badge: string;
          identity_enabled: boolean;
          identity_guild_id: string;
          tag: string;
        };
        collectibles: any;
        discriminator: string;
        display_name_styles: any;
        global_name: any;
        id: string;
        primary_guild: {
          badge: string;
          identity_enabled: boolean;
          identity_guild_id: string;
          tag: string;
        };
        public_flags: number;
        username: string;
      };
    };
    interaction_metadata: {
      authorizing_integration_owners: {
        "0": string;
      };
      command_type: number;
      id: string;
      name: string;
      type: number;
      user: {
        avatar: string;
        avatar_decoration_data: any;
        clan: {
          badge: string;
          identity_enabled: boolean;
          identity_guild_id: string;
          tag: string;
        };
        collectibles: any;
        discriminator: string;
        display_name_styles: any;
        global_name: any;
        id: string;
        primary_guild: {
          badge: string;
          identity_enabled: boolean;
          identity_guild_id: string;
          tag: string;
        };
        public_flags: number;
        username: string;
      };
    };
    mention_everyone: boolean;
    mention_roles: Array<any>;
    mentions: Array<any>;
    pinned: boolean;
    timestamp: string;
    tts: boolean;
    type: number;
    webhook_id: string;
  };
  token: string;
  type: number;
  version: number;
};